+++ src/Gate.IO.Api/Helpers/OptionsHelpers.cs
@@ -7,7 +7,7 @@ internal static class OptionsHelpers
         if (string.IsNullOrEmpty(symbol))
             throw new ArgumentException("Symbol is not provided");
 
-        if (!Regex.IsMatch(symbol, "^([A-Z|0-9]{2,})_([A-Z|0-9]{2,})$"))
+        if (!Regex.IsMatch(symbol, "^([A-Z0-9]{2,})_([A-Z0-9]{2,})$"))
             throw new ArgumentException($"{symbol} is not a valid symbol. Should be [BaseAsset]_[QuoteAsset], e.g. BTC_USDT");
     }
 
@@ -16,7 +16,7 @@ internal static class OptionsHelpers
         if (string.IsNullOrEmpty(symbol))
             throw new ArgumentException("Symbol is not provided");
 
-        if (!Regex.IsMatch(symbol, "^([A-Z|0-9]{2,})_([A-Z|0-9]{2,})-([0-9]{8})-([0-9])-([C|P])$"))
+        if (!Regex.IsMatch(symbol, "^([A-Z0-9]{2,})_([A-Z0-9]{2,})-([0-9]{8})-([0-9]+(?:\\.[0-9]+)?)-([CP])$"))
             throw new ArgumentException($"{symbol} is not a valid symbol. Should be [BaseAsset]_[QuoteAsset]-[Date]-[Number]-[C|P], e.g. BTC_USDT-20230310-30000-P");
     }
 }